# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Newsight API
  description: REST API documentation for the Newsight backend.
  contact:
    name: Newsight
  version: v1
servers:
- url: http://localhost:8080
  description: Local backend
- url: https://newsightkr.com
  description: Production
...
paths:
...
  /api/auth/find-password:
    post:
      tags:
      - Account Recovery
      summary: Issue temporary password
      operationId: findPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindPasswordRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FindPasswordResponseDto"
  /api/auth/find-password/:
    post:
      tags:
      - Account Recovery
      summary: Issue temporary password
      operationId: findPassword_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindPasswordRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FindPasswordResponseDto"
  /api/auth/find-id:
    post:
      tags:
      - Account Recovery
      summary: Find user ID
      operationId: findId
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindIdRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FindIdResponseDto"
  /api/auth/find-id/:
    post:
      tags:
      - Account Recovery
      summary: Find user ID
      operationId: findId_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindIdRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FindIdResponseDto"
  /api/auth/change-password:
    post:
      tags:
      - Account Password
      summary: Change password
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChangePasswordResponseDto"
      security:
      - bearerAuth: []
  /api/auth/change-password/:
    post:
      tags:
      - Account Password
      summary: Change password
      operationId: changePassword_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChangePasswordResponseDto"
      security:
      - bearerAuth: []
  /api/admins/users/withdraw:
    post:
      tags:
      - Admin Users
      summary: Withdraw user
      operationId: withdraw
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WithdrawResponseDto"
      security:
      - bearerAuth: []
  /api/admins/users/withdraw/:
    post:
      tags:
      - Admin Users
      summary: Withdraw user
      operationId: withdraw_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WithdrawResponseDto"
      security:
      - bearerAuth: []
  /api/admins/promote:
    post:
      tags:
      - Admin Users
      summary: Promote user to admin
      operationId: promote
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPromoteRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminPromoteResponseDto"
      security:
      - bearerAuth: []
  /api/admins/promote/:
    post:
      tags:
      - Admin Users
      summary: Promote user to admin
      operationId: promote_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPromoteRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminPromoteResponseDto"
      security:
      - bearerAuth: []
  /api/admins/demote/:
    post:
      tags:
      - Admin Users
      summary: Demote admin to user
      operationId: demote
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDemoteRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDemoteResponseDto"
      security:
      - bearerAuth: []
  /api/admins/demote:
    post:
      tags:
      - Admin Users
      summary: Demote admin to user
      operationId: demote_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDemoteRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDemoteResponseDto"
      security:
      - bearerAuth: []
...
  /api/admins/users/:
    get:
      tags:
      - Admin Users
      summary: List users
      operationId: listUsers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserListResponseDto"
      security:
      - bearerAuth: []
  /api/admins/users:
    get:
      tags:
      - Admin Users
      summary: List users
      operationId: listUsers_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserListResponseDto"
      security:
      - bearerAuth: []
...
  /api/admins/dashboard/login-logs/:
    get:
      tags:
      - Admin Dashboard
      summary: List login logs
      operationId: listAdminDashboardLoginLogs
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDashboardLoginLogsResponseDto"
      security:
      - bearerAuth: []
  /api/admins/dashboard/login-logs:
    get:
      tags:
      - Admin Dashboard
      summary: List login logs
      operationId: listAdminDashboardLoginLogs_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDashboardLoginLogsResponseDto"
      security:
      - bearerAuth: []
components:
  schemas:
...
    FindPasswordRequestDto:
      type: object
      properties:
        user_id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
      required:
      - email
      - name
      - user_id
    FindPasswordResponseDto:
      type: object
      properties:
        message:
          type: string
        temp_password:
          type: string
    FindIdRequestDto:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
      required:
      - email
      - name
    FindIdResponseDto:
      type: object
      properties:
        user_id:
          type: string
        message:
          type: string
    ChangePasswordRequestDto:
      type: object
      properties:
        current_password:
          type: string
          minLength: 1
        new_password:
          type: string
          minLength: 1
        new_password_confirm:
          type: string
          minLength: 1
      required:
      - current_password
      - new_password
      - new_password_confirm
    ChangePasswordResponseDto:
      type: object
      properties:
        message:
          type: string
        clear_refresh_cookie:
          type: boolean
    WithdrawRequestDto:
      type: object
      properties:
        user_seq:
          type: integer
          format: int64
      required:
      - user_seq
    WithdrawResponseDto:
      type: object
      properties:
        user_seq:
          type: integer
          format: int64
        deleted_at:
          type: string
          format: date-time
        acted_seq:
          type: integer
          format: int64
    AdminPromoteRequestDto:
      type: object
      properties:
        user_seq:
          type: integer
          format: int64
      required:
      - user_seq
    AdminPromoteResponseDto:
      type: object
      properties:
        user_seq:
          type: integer
          format: int64
        acted_seq:
          type: integer
          format: int64
        admin_level:
          type: string
        granted_at:
          type: string
          format: date-time
        message:
          type: string
    AdminDemoteRequestDto:
      type: object
      properties:
        user_seq:
          type: integer
          format: int64
      required:
      - user_seq
    AdminDemoteResponseDto:
      type: object
      properties:
        user_seq:
          type: integer
          format: int64
        acted_seq:
          type: integer
          format: int64
        demoted_at:
          type: string
          format: date-time
        message:
          type: string
...
    UserListItemDto:
      type: object
      properties:
        user_seq:
          type: integer
          format: int64
        user_id:
          type: string
        user_name:
          type: string
        grade_code:
          type: integer
          format: int32
        grade_name:
          type: string
        email:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
        last_login_at:
          type: string
          format: date-time
        joined_at:
          type: string
          format: date-time
        granted_at:
          type: string
          format: date-time
        password_changed_at:
          type: string
          format: date-time
    UserListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserListItemDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_count:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
        message:
          type: string
...
    AdminDashboardLoginLogsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LoginLogItemDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_count:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
    LoginLogItemDto:
      type: object
      properties:
        login_log_seq:
          type: integer
          format: int64
        input_id:
          type: string
        attempted_at:
          type: string
          format: date-time
        user_seq:
          type: integer
          format: int64
        is_success:
          type: boolean
        ip_address:
          type: string
        user_agent:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Use the access token returned by the login API.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT